Revert "Añadida implementación de UART para TMC"
