+++ tripteron_firmware/inc/hardware_manager/hardware_manager.h
@@ -7,7 +7,6 @@
 #include "i2c_manager_interface.h"
 #include "timer_manager_interface.h"
 #include "tmc_manager_interface.h"
-#include "uart_manager_interface.h"
 #include <memory>
 
 class HardwareManager : public IHardwareManager {
@@ -23,5 +22,4 @@ private:
     std::shared_ptr<ITmcManager> mTmcManager;
     std::shared_ptr<II2CManager> mI2CManager;
     std::shared_ptr<IEncoderManager> mEncoderManager;
-    std::shared_ptr<IUartManager> mUartManager;
 };
